Add/gradle and executable utils (#3)
Added utilities for reading gradle test logs, and some commonly used fake executables.

`GradleRunResult` receives the the logs of a gradle execution, and is able to break it in per-task logs, show the order of execution of these tasks, as well as compare execution order.

With a log for a ran task `TaskName` and its dependencies `dep1` and `dep2`.
```groovy
String log = "..." //gradle execution log
def result = new GradleRunResult(log)

//returns all tasks in execution order
result.orderedTasks

//execution order comparision, all of those are booleans
result["dep1"].wasExecutedBefore("dep2", ":taskName")
result["dep2"].wasExecutedBefore("taskName")
//task name can be either ':task' or just 'task'
result[":dep2"].wasExecutedAfter("dep1")
result[":taskName"].wasExecutedAfter(":dep1", "dep2")

//returns a string with logs for the given task
result["taskName"].taskLog 
```

`FakeExecutables` is a static functions class which include two methods: `runFirstParam` and `argsReflector`. 

`runFirstParam` returns a script `File` which runs the first param given to it, that must be a executable path, passing the remainder parameters to this executable. This is useful to mock "wrappers" like `mono`, for instance.

`argsReflector` returns a script file wrapped inside a `ArgsReflectorExecutable` object, which returns the parameters that this file was ran on, was well as its environment at execution time.  This can be used to mock executables that are too heavy or too complex to be tested using the real files, as the Unity3D executable, for instance.

``` groovy
//this creates a `ArgsReflectorExecutable` object
 def reflectorExec = FakeExecutables.argsReflector("filePath", exitStatus)

//executable file that should be used to run this object somewhere
 reflectorExec.executable

//returns the first execution result of this file in a given log
 ArgsReflectorExecutable.Result result = reflectorExec.firstResult(log)
 //you can also fetch all results inside a given log like this
 List< ArgsReflectorExecutable.Result> results = reflectorExec.allResults(log) 
 
//with an ArgsReflectorExecutable.Result object, you can get:
 result.args //List<String> of arguments passed to this executable
 result.envs //Map<Sring, String> representing the environment at the executable execution
```

Some extra, general utility functions were also added under `SpecUtils` and `GradleSpecUtils`

Co-authored-by: Joaquim Neto <joaquim.neto@wooga.net>