Description ===========
This patch adds snyk monitoring to the build pipeline.
It will hook itself into the check and publish stages.

The patch also sets a dependency helper plugin
`net.wooga.cve-dependency-resolution` which applies overrides
for dependencies with know fixes for security issues.

Along with the introduction of snyk I also upgraded/removed some
depdencies. Coveralls produces a huge load of errors even with the
latest version so I decided to remove it since we want/are moving
to sonarqube (It is unknown at this time if this dependency is actually better
or not).

Changes
=======

* ![ADD] `snyk` monitoring
* ![ADD] `net.wooga.snyk-wdk-java` snyk convention plugin
* ![ADD] `net.wogoa.cve-dependency-resolution` plugin
* ![REMOVE] coveralls plugin
* ![UPDATE] `com.google.guava:guava` to version `30.0-jre`
* ![UPDATE] `com.github.stefanbirkner:system-rules` to version `[1.18,2)`
* ![UPDATE] `junit:junit` to version `4.13.1`