fixes wrapValueBasedOnType not actually returning a provider when asked for it (#14)
Fixes issue on wrapValueBasedOnType where generated code for types Provider<RegularFile> and Provider<Directory> weren't actually returning the Provider, but instead the base types RegularFile and Directory. Also transformed wrapValueBasedOnType in a static method so helper functions on test code doesn't need to depend upon the IntegrationSpec class. As there were no other changes beside putting a static in front of the method, this is not a breaking change (unless someone is doing some deep reflection wizardry with it).

Also also, there is an inconsistency with the generated directory path in the same function when generating for type RegularFile/Directory vs. the corresponding providers. While the former generates with a path based on layout.projectDirectory, the later generates with a path based on layout.buildDirectory. Hadn't fixed it on this PR as it is a breaking change. It's possible to fix this without breaking existing behaviour, but it requires more work (or to deprecate the current function), so its better for it to be done in another PR.