Fix envNameFromProperty when numbers are part of the name
Description
===========

The helper method `envNameFromProperty` creates incorrect outputs
when the given propertyname contains numbers eg. `foo.bar22Bar`.
The expected result should be `FOO_BAR_22_BAR`.

I fixed the method with this patch and adjusted the tests. I also
made sure that the `toCamelCase` converts to lowercase before
splitting the values.

Note:
I know that `com.wooga.gradle:gradle-commons` also contains the
same method. So we could just bind to this implementation. But
I would urge to wait for a few more releases of this library to
not force incompatible upgrades if possible. This is the reason why
I pulled the same tests over here for now.

Changes
=======

* ![FIX] `envNameFromProperty` utility method
* ![FIX] `toCamelCase` utility method