Improve BatchmodeWrapper argument checks (#27)
Description
===========

The implementation for `containsArguments` was assuming that
the provided `List<String>` contains all arguments in order.
This is not very helpfull since we need to test if a given argument pair, or
multiple is passed to the tool. And we can't assume an oder since this
is implemenation detail. I switched the implementation to be more
in line with the check for given environment variables.
I renamed the old method to `containsAllArguments` to visualize the
differentce. I adjusted the wrapper script to print also an end marker
for the arguments/environment values so the logic can try to fetch
the whole section.

Changes
=======

* ![IMPROVE] `BatchmodeWrapper` argument checks